// types.ts
export interface Service {
  id: string;
  title: string;
  description: string;
  icon: string;
  image: string;
}

export interface Project {
  id: string;
  name: string;
  role: string;
  description: string;
  image: string;
}

export interface UpcomingProject {
  id: string;
  title: string;
  description: string;
}